+++ internal/fs/iswin_other.go
@@ -1,0 +1,8 @@
+// +build !js !wasm
+// +build !windows
+
+package fs
+
+func checkIfWindows() bool {
+	return false
+}