+++ internal/fs/iswin_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package fs
+
+func checkIfWindows() bool {
+	return true
+}